Added port_counters module to python tools
This is a merge of the individual node and switch packet stats scripts
used in the TestThroughputMulti evaluation code. This new version
retrieves port counters from a switch (via SNMP) and from a collection
of nodes (via SSH/ifconfig), stores them to JSON files and/or calculates
the difference from a pair of files. A mapping is defined to allow the
node stats to be associated with the appropriate switch port stats.