ExcaliburFrameDecoder now uses 'real' frame number in notifcations.
Instead of using the packet header frame number (which never gets reset
in current f/w), the frame decoder now puts the real frame number from
the subframe trailers into the frame ready notification passed to the
fileWriter.

Note that there is still a pathological case, where _both_ EOF subframe
packets are dropped so the subframe trailers are never seen, which could
cause the frame to be labelled with the header frame number. This will
be resolved in the upcoming 'farm mode' f/w release.