Remove config object parameter from controller configure call
The default configuration object built by command-line parsing in the
FrameReceiverApp class is now passed by reference to the constructor of
the FrameReceiverContoller. It is therefore unnecessary to pass to the
configure method. Also added a parameter to the config object to force
reconfiguration when passed in as an IPC message at startup. This will
be used to construct the reconfiguration dependency chain as the various
configuration steps are evaulated.