Add capability to IpcChannel to unbind from endpoint.
Allows IpcChannels to be unbound from bound endpoints. The class
stores a map of bound endpoints as ZeroMQ will resolve any wildcard
addresses in an endpoint URI passed to the bind() call and requires
that any unbind() call uses the resolved endpoint. Failure to do this
causes subsequent errors in socket operations.

It is worth being aware that ZeroMQ supports binding a socket to
multiple endpoints (each being associated with an underlying message
queue and transport-specific socket), so it is up to the application
using IpcChannel to either handle this or enforce a single bound
endpoint.