Allow RX thread to be reconfigured by control message changes.
As part of the client control implementation for the FR, it is sometimes
necessary to tear down and restart the RX thread due to changes in e.g.
the RX channel configuration. This exposed a serious shortcoming in the
ZeroMQ PAIR socket implementation - if a thread at the end of the PAIR
channel disconnects, termiantes and is restarted, any subsequent send()
calls from EITHER end of the socket channel will hang. The RX channel
has therefore been switched to a DEALER-ROUTER pattern, including the
ability for the RX channel to identify and use a unique identity. It
also seems that DEALER-ROUTER doesn't like new endpoints in threads to
use the same identity as a previous endpoint.

Here endeth today's lesson in ZeroMQ arcana. It works well when it works
but there's a lot you have to find out for yourself by trial and
error...