Add conditional reconfguration of frame decoder and buffer manager.
The conditional reconfiguration on reception of a configuration message
by the FrameController now includes the frame decoder and the shared
buffer manager. This includes correct management of the reconfiguration
dependency chain and, where necessary, early termination of the RX
thread to allow existing decoder and buffer manager instances to be
cleaned up.