First pass at handling config request message in FrameReceiver.
All configuration parameters are built into the reply. Need to consider
if decoder params should come from config tree stored in the app
controller or by making explicit call to decoder function (like FP).
At present, any decoder params not explicitly set at startup or in
control channel messages (and therefore getting default values in
the decoder) won't be seen as they are not stored in the controller
config cache. Not sure if this is a real issue...