Prevent frameReceiver controller from blocking when RX thread startup fails.
The RX thread startup error handling has been improved to prevent the main
thread controller from blocking when an initialisation error occurs (e.g.
failing to bind to RX ports). The RxThread start() method was calling join()
when the thread init fails, without allowing the thread to terminate. This
prevented reporting of the RX thread init error and the startup of the main
controller reactor, rendering the application unresponsive.

Fixes #104