Fix RX thread startup wait loop.
The FrameReceiverRxThread::start() method has a wait loop which blocks
until the thread is running and initialised. When compiled as a CMake
release build, i.e. with -O3 optimisation, the loop was being optimised
in such a way that the polling in the loop did not detect the thread
state change. This has been fixed by modifying the loop to have a
bounded timeout and a short sleep each on pass.

Resolves #248