Update ZeroMQ CMake rules to explicitly resolve and check version.
The version of cmake/FindZeroMQ.cmake in use did not attempt to resolve
the ZeroMQ version from the identified installation. There is also a
'feature' in CMake that requires the case of the rules file to match
that of the package being resolved, hence the file name is changed to
FindZEROMQ.cmake.