Store sockets as a dictionary by endpoint.
"data_sockets" was previously a list which made an assumption that all FPs in
the system would have an increasing rank starting at 0 and going up by 1.  This
change allows all the FPs in the system to use the meta writer because the
endpoint of a FP is unique, whereas the rank is not guaranteed to be unique.