Fix elapsed time calculation in live view rate settiing.
The elapsed time calcuation relies on truncating a posix_time duration
to milliseconds as an int_32, which can fail with an incorrect value
on the initial frame, or if the frame processor has been idling for
long periods. To resolve this, the time since last frame variable
is now stored and compared as a time_duration.