Update to use improved h5py string support
This allows users of StringHDF5Dataset to write data as fixed-length strings
or as variable-length objects depending on whether `length` is passed. It
handles various conversions including non-string types that can be converted
to `bytes`.

The `encoding` parameter of `string_dtype` is exposed using the underlying
default, `"utf-8"`. This can be overridden to encode as `ascii`.