Refactor FrameProcessorController::handleCtrlChannel()
Replace if-else case with switch case. Switch case uses a jump-table to
the exact matching case, if-else uses a compare-jump-to-label semantics.
The if-else case will go through all the cases to find it's match, the
switch-case uses a Jump target to match the exact case. From the disassembly of the instructions.